+++ simple_dvm/type_ids_parser.c
@@ -10,7 +10,7 @@ void parse_type_ids(DexFileFormat *dex, unsigned char *buf, int offset)
 {
     int i = 0;
     if (is_verbose() > 3)
-        printf("parse type ids offset = %04x\n", offset + sizeof(DexHeader));
+        printf("parse type ids offset = %04x\n", (uint)(offset + sizeof(DexHeader)));
     dex->type_id_item = malloc(
                             sizeof(type_id_item) * dex->header.typeIdsSize);
 
@@ -46,7 +46,7 @@ void parse_proto_ids(DexFileFormat *dex, unsigned char *buf, int offset)
     volatile int i = 0, j = 0;
     int idx = 0;
     if (is_verbose() > 3)
-        printf("parse proto ids offset = %04x\n", offset + sizeof(DexHeader));
+        printf("parse proto ids offset = %04x\n", (uint)(offset + sizeof(DexHeader)));
     dex->proto_id_item = malloc(
                              sizeof(proto_id_item) * dex->header.protoIdsSize);
 
@@ -117,7 +117,7 @@ void parse_field_ids(DexFileFormat *dex, unsigned char *buf, int offset)
 {
     int i;
     if (is_verbose() > 3)
-        printf("parse feild ids offset = %04x\n", offset + sizeof(DexHeader));
+        printf("parse feild ids offset = %04x\n", (uint)(offset + sizeof(DexHeader)));
     dex->field_id_item = malloc(sizeof(field_id_item) * dex->header.fieldIdsSize);
 
     if (is_verbose() > 3)
